rename all your scripts with standard name conversion
